Separate classes into their own files
