+++ src/System.Net.Http/src/System.Net.Http.csproj
@@ -122,19 +122,28 @@
     <Compile Include="System\Net\Http\Managed\AuthenticationHandler.cs" />
     <Compile Include="System\Net\Http\Managed\AutoRedirectHandler.cs" />
     <Compile Include="System\Net\Http\Managed\BasicAuthenticationHelper.cs" />
+    <Compile Include="System\Net\Http\Managed\ChunkedEncodingReadStream.cs" />
+    <Compile Include="System\Net\Http\Managed\ChunkedEncodingWriteStream.cs" />
     <Compile Include="System\Net\Http\Managed\ConnectHelper.cs" />
+    <Compile Include="System\Net\Http\Managed\ConnectionCloseReadStream.cs" />
+    <Compile Include="System\Net\Http\Managed\ContentLengthReadStream.cs" />
+    <Compile Include="System\Net\Http\Managed\ContentLengthWriteStream.cs" />
     <Compile Include="System\Net\Http\Managed\CookieHandler.cs" />
     <Compile Include="System\Net\Http\Managed\DecompressionHandler.cs" />
+    <Compile Include="System\Net\Http\Managed\EmptyReadStream.cs" />
     <Compile Include="System\Net\Http\Managed\HttpConnection.cs" />
+    <Compile Include="System\Net\Http\Managed\HttpConnectionContent.cs" />
     <Compile Include="System\Net\Http\Managed\HttpConnectionHandler.cs" />
     <Compile Include="System\Net\Http\Managed\HttpConnectionKey.cs" />
     <Compile Include="System\Net\Http\Managed\HttpConnectionPool.cs" />
     <Compile Include="System\Net\Http\Managed\HttpConnectionPools.cs" />
     <Compile Include="System\Net\Http\Managed\HttpConnectionSettings.cs" />
     <Compile Include="System\Net\Http\Managed\HttpContentReadStream.cs" />
+    <Compile Include="System\Net\Http\Managed\HttpContentStream.cs" />
     <Compile Include="System\Net\Http\Managed\HttpContentWriteStream.cs" />
     <Compile Include="System\Net\Http\Managed\HttpProxyConnectionHandler.cs" />
     <Compile Include="System\Net\Http\Managed\ManagedHandler.cs" />
+    <Compile Include="System\Net\Http\Managed\ValueStringBuilder.cs" />
   </ItemGroup>
   <!-- Common, except for Windows net46 build -->
   <ItemGroup Condition="'$(TargetGroup)' != 'net46'  And '$(TargetGroup)' != 'netfx'">
@@ -422,6 +431,7 @@
     <Reference Include="System.Diagnostics.Tools" />
     <Reference Include="System.Diagnostics.Tracing" />
     <Reference Include="System.IO.Compression" />
+    <Reference Include="System.Memory" />
     <Reference Include="System.Net.Primitives" />
     <Reference Include="System.Net.Security" />
     <Reference Include="System.Net.Sockets" />
