+++ src/System.Net.Http/src/System/Net/Http/Managed/HttpConnection.cs
@@ -13,7 +13,7 @@ using System.Threading.Tasks;
 
 namespace System.Net.Http
 {
-    internal sealed class HttpConnection : IDisposable
+    internal sealed partial class HttpConnection : IDisposable
     {
         private const int InitialReadBufferSize =
 #if DEBUG
@@ -47,435 +47,6 @@ namespace System.Net.Http
         private bool _connectionClose;      // Connection: close was seen on last response
 
         private bool _disposed;
-
-        private sealed class HttpConnectionContent : HttpContent
-        {
-            private readonly CancellationToken _cancellationToken;
-            private HttpContentReadStream _stream;
-
-            public HttpConnectionContent(CancellationToken cancellationToken)
-            {
-                _cancellationToken = cancellationToken;
-            }
-
-            public void SetStream(HttpContentReadStream stream)
-            {
-                Debug.Assert(stream != null);
-                Debug.Assert(stream.CanRead);
-
-                _stream = stream;
-            }
-
-            private HttpContentReadStream ConsumeStream()
-            {
-                if (_stream == null)
-                {
-                    throw new InvalidOperationException("content already consumed");
-                }
-
-                HttpContentReadStream stream = _stream;
-                _stream = null;
-                return stream;
-            }
-
-            protected override async Task SerializeToStreamAsync(Stream stream, TransportContext context)
-            {
-                Debug.Assert(stream != null);
-
-                using (HttpContentReadStream contentStream = ConsumeStream())
-                {
-                    const int BufferSize = 8192;
-                    await contentStream.CopyToAsync(stream, BufferSize, _cancellationToken).ConfigureAwait(false);
-                }
-            }
-
-            protected internal override bool TryComputeLength(out long length)
-            {
-                length = 0;
-                return false;
-            }
-
-            protected override Task<Stream> CreateContentReadStreamAsync() =>
-                Task.FromResult<Stream>(ConsumeStream());
-
-            internal override Stream TryCreateContentReadStream() =>
-                ConsumeStream();
-
-            protected override void Dispose(bool disposing)
-            {
-                if (disposing)
-                {
-                    if (_stream != null)
-                    {
-                        _stream.Dispose();
-                        _stream = null;
-                    }
-                }
-
-                base.Dispose(disposing);
-            }
-        }
-
-        private sealed class EmptyReadStream : HttpContentReadStream
-        {
-            internal static EmptyReadStream Instance { get; } = new EmptyReadStream();
-
-            private readonly static Task<int> s_zeroTask = Task.FromResult(0);
-
-            private EmptyReadStream() : base(null) { }
-
-            public override Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-            {
-                ValidateBufferArgs(buffer, offset, count);
-                return s_zeroTask;
-            }
-        }
-
-
-        private sealed class ContentLengthReadStream : HttpContentReadStream
-        {
-            private long _contentBytesRemaining;
-
-            public ContentLengthReadStream(HttpConnection connection, long contentLength)
-                : base(connection)
-            {
-                if (contentLength == 0)
-                {
-                    _connection = null;
-                    _contentBytesRemaining = 0;
-                    connection.ReturnConnectionToPool();
-                }
-                else
-                {
-                    _contentBytesRemaining = contentLength;
-                }
-            }
-
-            public override async Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-            {
-                ValidateBufferArgs(buffer, offset, count);
-
-                if (_connection == null || count == 0)
-                {
-                    // Response body fully consumed or the caller didn't ask for any data
-                    return 0;
-                }
-
-                Debug.Assert(_contentBytesRemaining > 0);
-
-                count = (int)Math.Min(count, _contentBytesRemaining);
-
-                int bytesRead = await _connection.ReadAsync(buffer, offset, count, cancellationToken).ConfigureAwait(false);
-
-                if (bytesRead == 0)
-                {
-                    // Unexpected end of response stream
-                    throw new IOException("Unexpected end of content stream");
-                }
-
-                Debug.Assert(bytesRead <= _contentBytesRemaining);
-                _contentBytesRemaining -= bytesRead;
-
-                if (_contentBytesRemaining == 0)
-                {
-                    // End of response body
-                    _connection.ReturnConnectionToPool();
-                    _connection = null;
-                }
-
-                return bytesRead;
-            }
-
-            public override async Task CopyToAsync(Stream destination, int bufferSize, CancellationToken cancellationToken)
-            {
-                if (destination == null)
-                {
-                    throw new ArgumentNullException(nameof(destination));
-                }
-
-                if (_connection == null)
-                {
-                    // Response body fully consumed
-                    return;
-                }
-
-                await _connection.CopyChunkToAsync(destination, _contentBytesRemaining, cancellationToken).ConfigureAwait(false);
-
-                _contentBytesRemaining = 0;
-                _connection.ReturnConnectionToPool();
-                _connection = null;
-            }
-        }
-
-        private sealed class ChunkedEncodingReadStream : HttpContentReadStream
-        {
-            private int _chunkBytesRemaining;
-
-            public ChunkedEncodingReadStream(HttpConnection connection)
-                : base(connection)
-            {
-                _chunkBytesRemaining = 0;
-            }
-
-            private async Task<bool> TryGetNextChunk(CancellationToken cancellationToken)
-            {
-                Debug.Assert(_chunkBytesRemaining == 0);
-
-                // Start of chunk, read chunk size.
-                int chunkSize = ParseHexSize(await _connection.ReadNextLineAsync(cancellationToken).ConfigureAwait(false));
-                _chunkBytesRemaining = chunkSize;
-
-                if (chunkSize > 0)
-                {
-                    return true;
-                }
-
-                // Indicates end of response body. We expect final CRLF after this.
-                if ((await _connection.ReadNextLineAsync(cancellationToken).ConfigureAwait(false)).Count != 2) // \r\n
-                {
-                    throw new IOException("missing final CRLF for chunked encoding");
-                }
-
-                _connection.ReturnConnectionToPool();
-                _connection = null;
-                return false;
-            }
-
-            private int ParseHexSize(ArraySegment<byte> line)
-            {
-                // TODO #21452: Handle overflow of size
-                long size = 0;
-                for (int i = 0; i < line.Count; i++)
-                {
-                    char c = (char)line[i];
-                    if ((uint)(c - '0') <= '9' - '0')
-                    {
-                        size = size * 16 + (c - '0');
-                    }
-                    else if ((uint)(c - 'a') <= ('f' - 'a'))
-                    {
-                        size = size * 16 + (c - 'a' + 10);
-                    }
-                    else if ((uint)(c - 'A') <= ('F' - 'A'))
-                    {
-                        size = size * 16 + (c - 'A' + 10);
-                    }
-                    else if (c == '\r')
-                    {
-                        break;
-                    }
-                    else
-                    {
-                        throw new IOException("Invalid chunk size in response stream");
-                    }
-                }
-                return (int)size;
-            }
-
-            private async Task ConsumeChunkBytes(int bytesConsumed, CancellationToken cancellationToken)
-            {
-                Debug.Assert(bytesConsumed <= _chunkBytesRemaining);
-                _chunkBytesRemaining -= bytesConsumed;
-                if (_chunkBytesRemaining == 0 &&
-                    (await _connection.ReadNextLineAsync(cancellationToken).ConfigureAwait(false)).Count != 2)
-                {
-                    throw new IOException("missing CRLF for end of chunk");
-                }
-            }
-
-            public override async Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-            {
-                ValidateBufferArgs(buffer, offset, count);
-
-                if (_connection == null || count == 0)
-                {
-                    // Response body fully consumed or the caller didn't ask for any data
-                    return 0;
-                }
-
-                if (_chunkBytesRemaining == 0)
-                {
-                    if (!await TryGetNextChunk(cancellationToken).ConfigureAwait(false))
-                    {
-                        // End of response body
-                        return 0;
-                    }
-                }
-
-                count = Math.Min(count, _chunkBytesRemaining);
-
-                int bytesRead = await _connection.ReadAsync(buffer, offset, count, cancellationToken).ConfigureAwait(false);
-
-                if (bytesRead == 0)
-                {
-                    // Unexpected end of response stream
-                    throw new IOException("Unexpected end of content stream while processing chunked response body");
-                }
-
-                await ConsumeChunkBytes(bytesRead, cancellationToken).ConfigureAwait(false);
-
-                return bytesRead;
-            }
-
-            public override async Task CopyToAsync(Stream destination, int bufferSize, CancellationToken cancellationToken)
-            {
-                if (destination == null)
-                {
-                    throw new ArgumentNullException(nameof(destination));
-                }
-
-                if (_connection == null)
-                {
-                    // Response body fully consumed
-                    return;
-                }
-
-                if (_chunkBytesRemaining > 0)
-                {
-                    await _connection.CopyChunkToAsync(destination, _chunkBytesRemaining, cancellationToken).ConfigureAwait(false);
-                    await ConsumeChunkBytes(_chunkBytesRemaining, cancellationToken).ConfigureAwait(false);
-                }
-
-                while (await TryGetNextChunk(cancellationToken).ConfigureAwait(false))
-                {
-                    await _connection.CopyChunkToAsync(destination, _chunkBytesRemaining, cancellationToken).ConfigureAwait(false);
-                    await ConsumeChunkBytes(_chunkBytesRemaining, cancellationToken).ConfigureAwait(false);
-                }
-            }
-        }
-
-        private sealed class ConnectionCloseReadStream : HttpContentReadStream
-        {
-            public ConnectionCloseReadStream(HttpConnection connection)
-                : base(connection)
-            {
-            }
-
-            public override async Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-            {
-                ValidateBufferArgs(buffer, offset, count);
-
-                if (_connection == null || count == 0)
-                {
-                    // Response body fully consumed or the caller didn't ask for any data
-                    return 0;
-                }
-
-                int bytesRead = await _connection.ReadAsync(buffer, offset, count, cancellationToken).ConfigureAwait(false);
-
-                if (bytesRead == 0)
-                {
-                    // We cannot reuse this connection, so close it.
-                    _connection.Dispose();
-                    _connection = null;
-                    return 0;
-                }
-
-                return bytesRead;
-            }
-
-            public override async Task CopyToAsync(Stream destination, int bufferSize, CancellationToken cancellationToken)
-            {
-                if (destination == null)
-                {
-                    throw new ArgumentNullException(nameof(destination));
-                }
-
-                if (_connection == null)
-                {
-                    // Response body fully consumed
-                    return;
-                }
-
-                await _connection.CopyToAsync(destination, cancellationToken).ConfigureAwait(false);
-
-                // We cannot reuse this connection, so close it.
-                _connection.Dispose();
-                _connection = null;
-            }
-        }
-
-        private sealed class ChunkedEncodingWriteStream : HttpContentWriteStream
-        {
-            public ChunkedEncodingWriteStream(HttpConnection connection)
-                : base (connection)
-            {
-            }
-
-            public override async Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-            {
-                ValidateBufferArgs(buffer, offset, count);
-
-                // Don't write if nothing was given
-                // Especially since we don't want to accidentally send a 0 chunk, which would indicate end of body
-                if (count == 0)
-                {
-                    return;
-                }
-
-                // Write chunk length -- hex representation of count
-                bool digitWritten = false;
-                for (int i = 7; i >= 0; i--)
-                {
-                    int shift = i * 4;
-                    int mask = 0xF << shift;
-                    int digit = (count & mask) >> shift;
-                    if (digitWritten || digit != 0)
-                    {
-                        await _connection.WriteByteAsync((byte)(digit < 10 ? '0' + digit : 'A' + digit - 10), cancellationToken).ConfigureAwait(false);
-                        digitWritten = true;
-                    }
-                }
-
-                await _connection.WriteTwoBytesAsync((byte)'\r', (byte)'\n', cancellationToken).ConfigureAwait(false);
-
-                // Write chunk contents
-                await _connection.WriteAsync(buffer, offset, count, cancellationToken).ConfigureAwait(false);
-                await _connection.WriteTwoBytesAsync((byte)'\r', (byte)'\n', cancellationToken).ConfigureAwait(false);
-            }
-
-            public override Task FlushAsync(CancellationToken cancellationToken)
-            {
-                return _connection.FlushAsync(cancellationToken);
-            }
-
-            public override async Task FinishAsync(CancellationToken cancellationToken)
-            {
-                // Send 0 byte chunk to indicate end
-                await _connection.WriteByteAsync((byte)'0', cancellationToken).ConfigureAwait(false);
-                await _connection.WriteTwoBytesAsync((byte)'\r', (byte)'\n', cancellationToken).ConfigureAwait(false);
-
-                // Send final _CRLF
-                await _connection.WriteTwoBytesAsync((byte)'\r', (byte)'\n', cancellationToken).ConfigureAwait(false);
-
-                _connection = null;
-            }
-        }
-
-        public sealed class ContentLengthWriteStream : HttpContentWriteStream
-        {
-            public ContentLengthWriteStream(HttpConnection connection)
-                : base(connection)
-            {
-            }
-
-            public override Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-            {
-                ValidateBufferArgs(buffer, offset, count);
-                return _connection.WriteAsync(buffer, offset, count, cancellationToken);
-            }
-
-            public override Task FlushAsync(CancellationToken cancellationToken)
-            {
-                return _connection.FlushAsync(cancellationToken);
-            }
-
-            public override Task FinishAsync(CancellationToken cancellationToken)
-            {
-                _connection = null;
-                return Task.CompletedTask;
-            }
-        }
 
         public HttpConnection(
             HttpConnectionPool pool,
@@ -1270,33 +841,6 @@ namespace System.Net.Http
 
             // We're not putting the connection back in the pool. Dispose it.
             Dispose();
-        }
-
-        private struct ValueStringBuilder
-        {
-            public char[] Chars;
-            public int Length;
-
-            public ValueStringBuilder(int initialCapacity)
-            {
-                Chars = new char[initialCapacity];
-                Length = 0;
-            }
-
-            public void Append(char c)
-            {
-                if (Length == Chars.Length)
-                {
-                    Grow();
-                }
-                Chars[Length++] = c;
-            }
-
-            private void Grow() => Array.Resize(ref Chars, Chars.Length * 2);
-
-            public void Clear() => Length = 0;
-
-            public override string ToString() => new string(Chars, 0, Length);
         }
     }
 }
