+++ src/System.Net.Http/src/System/Net/Http/Managed/HttpContentReadStream.cs
@@ -37,13 +37,10 @@ namespace System.Net.Http
         public override void Write(byte[] buffer, int offset, int count) => throw new NotSupportedException();
 
         public override int Read(byte[] buffer, int offset, int count) =>
-            ReadAsync(buffer, offset, count, CancellationToken.None).Result;
+            ReadAsync(buffer, offset, count, CancellationToken.None).GetAwaiter().GetResult();
 
-        // TODO #21452: Restore this once we address unit test build's target
-        //public override void CopyTo(Stream destination, int bufferSize)
-        //{
-        //    CopyToAsync(destination, bufferSize, CancellationToken.None).Wait();
-        //}
+        public override void CopyTo(Stream destination, int bufferSize) =>
+            CopyToAsync(destination, bufferSize, CancellationToken.None).GetAwaiter().GetResult();
 
         protected override void Dispose(bool disposing)
         {
